Move cairo 1 fixtures to proper dir
