+++ starknet_py/tests/e2e/contract_interaction/v1_interaction_test.py
@@ -5,7 +5,7 @@ import pytest
 from starknet_py.cairo.felt import decode_shortstring, encode_shortstring
 from starknet_py.contract import Contract
 from starknet_py.tests.e2e.fixtures.constants import MAX_FEE
-from starknet_py.tests.e2e.fixtures.contracts import deploy_v1_contract
+from starknet_py.tests.e2e.fixtures.contracts_v1 import deploy_v1_contract
 
 # TODO (#1219): investigate why some of these tests fails for contracts_compiled_v1
